Rename & update function to reduceErrors
